Fix bugs, merge differences

+++ service/src/routes/invoice.ts
@@ -1,12 +1,15 @@
 import express from 'express'
-import axios from 'axios'
-import * as InvoiceController from "../controllers/invoice"
-import { IInvoice } from '../models/invoice';
+
+import * as InvoiceController from "../api/invoice"
+import { validateBearerToken } from '../models/user';
 
 const router = express.Router();
 
 // Get a list of invoices
 router.get("", async (req, res) => {
+    if (!validateBearerToken(req.headers.authorization, res))
+        return;
+
     InvoiceController.getInvoices(req.query)
         .then(result => res.status(200).send(result))
         .catch(err => res.status(400).send(err));
@@ -14,6 +17,9 @@ router.get("", async (req, res) => {
 
 // Get a list of invoices lines
 router.get("/invoiceLines", async (req, res) => {
+    if (!validateBearerToken(req.headers.authorization, res))
+        return;
+
     InvoiceController.getInvoiceLines(req.query)
         .then(result => res.status(200).send(result))
         .catch(err => res.status(400).send(err));
@@ -21,6 +27,9 @@ router.get("/invoiceLines", async (req, res) => {
 
 // Get an invoice by id
 router.get("/:invoiceId", async (req, res) => {
+    if (!validateBearerToken(req.headers.authorization, res))
+        return;
+
     InvoiceController.getInvoiceById(req.params.invoiceId)
         .then(result => res.status(200).send(result))
         .catch(err => res.status(400).send(err));
@@ -28,6 +37,9 @@ router.get("/:invoiceId", async (req, res) => {
 
 // Get an invoice line by id
 router.get("/invoiceLines/:invoiceLineId", async (req, res) => {
+    if (!validateBearerToken(req.headers.authorization, res))
+        return;
+
     InvoiceController.getInvoiceLineById(req.params.invoiceLineId).
         then(result => res.status(200).send(result))
         .catch(err => res.status(400).send(err));
@@ -35,6 +47,9 @@ router.get("/invoiceLines/:invoiceLineId", async (req, res) => {
 
 // Delete an invoice
 router.delete("/delete/:invoiceId", async (req, res) => {
+    if (!validateBearerToken(req.headers.authorization, res))
+        return;
+
     InvoiceController.deleteInvoice(req.params.invoiceId)
         .then(result => res.status(200).send(result))
         .catch(err => res.status(400).send(err));
@@ -42,6 +57,9 @@ router.delete("/delete/:invoiceId", async (req, res) => {
 
 // Delete an invoice line
 router.delete("/invoiceLines/delete/:invoiceLineId", async (req, res) => {
+    if (!validateBearerToken(req.headers.authorization, res))
+        return;
+
     InvoiceController.deleteInvoiceLine(req.params.invoiceLineId)
         .then(result => res.status(200).send(result))
         .catch(err => res.status(400).send(err));
@@ -49,12 +67,19 @@ router.delete("/invoiceLines/delete/:invoiceLineId", async (req, res) => {
 
 // Create an invoice
 router.post("/create", async (req, res) => {
+    if (!validateBearerToken(req.headers.authorization, res))
+        return;
+
     InvoiceController.createInvoice(req.body.invoiceData)
         .then(result => res.status(200).send(result))
         .catch(err => res.status(400).send(err));
 });
+
 // Update an invoice
 router.put("/update", async (req, res) => {
+    if (!validateBearerToken(req.headers.authorization, res))
+        return;
+
     InvoiceController.updateInvoice(req.body.id, req.body.newValues)
         .then(result => res.status(200).send(result))
         .catch(err => res.status(400).send(err));
@@ -62,10 +87,12 @@ router.put("/update", async (req, res) => {
 
 // Update invoice line
 router.put("/invoiceLines/update", async (req, res) => {
+    if (!validateBearerToken(req.headers.authorization, res))
+        return;
+
     InvoiceController.updateInvoiceLine(req.body.id, req.body.newValues)
         .then(result => res.status(200).send(result))
         .catch(err => res.status(400).send(err));
 });
-
 
 export default router;
